Conversion to a Montage app
I moved all files related to the chrome extension into the "extension" directory and placed the demo into the "example" directory.

This change just packages the existing code into a Montage app but it doesn't break down the elements, created at draw cycle, into components.

I believe that changing from the sendMessage API to a long-lived connection can improve communication from the page to the content-script.

The Main component is really tied into the chrome extension API, we should probably move the chrome extension calls into another layer/object.

We should also probably create a component for what is currently Main so we can move it around easier in the future if we need to.

A word of concern about "script-self 'unsafe-eval'": we can use it for development but I recommend removing it and use a mopped version for production.